Moved interfaces to respective sub-namespaces
Tests amended
